+++ lib/linter.dart
@@ -56,9 +56,9 @@ String _todosPrefPath = '${pluginId}.showTodos';
 String _filterUnnamedLibraryWarningsPath = '${pluginId}.x_filterUnnamedLibraryWarnings';
 String _filterCompiledToJSWarningsPath = '${pluginId}.x_filterCompiledToJSWarnings';
 
-bool _shouldShowInfoMessages() => atom.config.get(_infosPrefPath);
-bool _shouldShowTodosMessages() => atom.config.get(_todosPrefPath);
+bool _shouldShowInfoMessages() => atom.config.getValue(_infosPrefPath);
+bool _shouldShowTodosMessages() => atom.config.getValue(_todosPrefPath);
 bool _shouldFilterUnnamedLibraryWarnings() =>
-    atom.config.get(_filterUnnamedLibraryWarningsPath);
+    atom.config.getValue(_filterUnnamedLibraryWarningsPath);
 bool _shouldFilterCompiledToJSWarnings() =>
-    atom.config.get(_filterCompiledToJSWarningsPath);
+    atom.config.getValue(_filterCompiledToJSWarningsPath);